- Code cleanup... (less warnings under VC compiles) - Walk delay only takes effect if attack did damage. - Removed magnum break's icon.
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@4070 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/map/npc.c
@@ -2051,7 +2051,7 @@ static int npc_parse_function (char *w1, char *w2, char *w3, char *w4, char *fir
 		npc_parse_script_line(line,&curly_count,*lines);
 		if (feof(fp))
 			break;
-		if (strlen(srcbuf)+strlen(line)+1 >= srcsize) {
+		if (strlen(srcbuf)+strlen(line)+1 >= (unsigned int)srcsize) {
 			srcsize += 65536;
 			srcbuf = (char *)aRealloc(srcbuf, srcsize);
 			memset(srcbuf + srcsize - 65536, '\0', 65536);
